refactor(infra): Переименовать классы (#6):
- `WeatherCodeTranslator` -> `OpenMeteoWeatherCodeTranslator`;
- `FetchTimeProvider` -> `SystemClock`.
