--- Nubrio.Infrastructure/Services/FetchTimeProvider.cs
@@ -1,8 +1,0 @@
-using Nubrio.Application.Interfaces;
-
-namespace Nubrio.Infrastructure.Services;
-
-public class FetchTimeProvider : IClock
-{
-    public DateTimeOffset UtcNow => DateTimeOffset.Now;
-}
