+++ Nubrio.Infrastructure/Services/SystemClock.cs
@@ -1,0 +1,8 @@
+using Nubrio.Application.Interfaces;
+
+namespace Nubrio.Infrastructure.Services;
+
+public class SystemClock : IClock
+{
+    public DateTimeOffset UtcNow => DateTimeOffset.UtcNow;
+}
